4.21 delete with valid token

+++ controllers/blogs.js
@@ -8,14 +8,6 @@ blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
   })
-
-// const getTokenFrom = request => {
-//   const authorization = request.get('authorization')
-//   if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-//     return authorization.substring(7)
-//   }
-//   return null
-// }
   
 blogsRouter.post('/', async (request, response, next) => { 
   const body = request.body
@@ -24,7 +16,6 @@ blogsRouter.post('/', async (request, response, next) => {
     return response.status(400).end()
   }
 
-  //const token = getTokenFrom(request)
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!request.token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
@@ -52,12 +43,26 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
   blogsRouter.delete('/:id', async (request, response, next) => {
-    try{
-      await Blog.findByIdAndDelete(request.params.id)
-      response.status(204).end()
-    } catch(exception) {
-      next(exception)
+    const blog = await Blog.findById(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: "token missing or invalid" });
     }
+    const user = await User.findById(decodedToken.id);
+
+   if (blog.user.toString() === user._id.toString()) {
+    try {
+      await Blog.findByIdAndDelete(request.params.id);
+      response.status(204).end();
+    } catch (exception) {
+      next(exception);
+    }
+   } else {
+    return response.status(401).json({ error: 'unauthorized user' })
+   }
+
+    
   })
 
   blogsRouter.put('/:id', async (request, response, next) => {
